Make sending reports at shutdown configurable
The new Android Runtime, default as of Android 5.0, does not allow
spawning of threads during shutdown, and doing so raises
an exception:

  java.lang.InternalError: Thread starting during runtime shutdown

For reasons I don't fully understand, it appears the the shutdown
begins before ACRA even re-raises the exception. So the fact that
ACRA waits on the SendWorker is insufficient to prevent this issue
from occurring. For example, see issue #136, which in theory should
not occur.

This option allows ACRA to be configured to terminate immediately
after persisting the report to disk. The report is then sent
(per usual) when the application is restarted.