Prevent potential race-condition when sending reports at startup
In theory, the SendWorker() can start sending reports before
mReportSenders has been properly set. This is extremely unlikely
in practice due to the overhead of spawning a new thread and the
amount of code that is executed by the SendWorker before it needs
to access mReportSenders.

But we can prevent the race from occuring in theory as well as in
practice by deferring calling checkReportsOnApplicationStart()
till after mReportSenders has been set.

https://github.com/ACRA/acra/issues/200