Only collect initial configuration if we're going to report it
Using reflection to collect the initial device configuration is
fairly expensive and impacts application startup time, especially
on older devices. Profiling on a Moto G showed the
ConfigurationCollector taking 70% of the ACRA.init() CPU time.

The ConfigurationCollector has been converted to a plain object.
This means that if both INITIAL_CONFIGURATION and CRASH_CONFIGURATION
are being reported that the previously static initializer code will
now be run twice, but this is offset by the fact that we no longer
hold this data around in memory on the off-chance that we'll use it.