Put only `auto-service-annotations` in classpath, not `auto-service`.
Compare https://github.com/ACRA/acra/issues/738, which discussed a
change in documentation for _users'_ builds. This PR changes how ACRA
itself depends on AutoService.

This keeps the Auto-Service annotation processor, as well as
dependencies like Auto-Common and Guava, out of the runtime classpath.

By keeping Guava in particular out of the runtime classpath, we reduce
the chances that users encounter problems with Guava's strange
dependency structure, as discussed in
https://issuetracker.google.com/issues/457186394#comment5.